construction-estimator/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
reportlab==4.0.7
python-jose==3.3.0
bcrypt==3.2.2
python-dotenv==1.0.0
stripe==7.10.0
